Move extensions.go to webauthntypes/
