+++ lib/auth/webauthntypes/webauthn_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package webauthn_test
+package webauthntypes_test
 
 import (
 	"encoding/base64"
